commit final eu espero

+++ src/main/java/io/github/eliasfelipedasilva/domain/entity/Vaga.java
@@ -11,7 +11,7 @@ public class Vaga {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_vaga")
-    private Long  id_vaga;
+    private Integer  id_vaga;
 
     @Column(name = "cargo_vaga")
     private String cargo_vaga;
@@ -32,11 +32,11 @@ public class Vaga {
     private String status_vaga;
 
 
-    public Long getId_vaga() {
+    public Integer getId_vaga() {
         return id_vaga;
     }
 
-    public void setId_vaga(Long id_vaga) {
+    public void setId_vaga(Integer id_vaga) {
         this.id_vaga = id_vaga;
     }
